Usage of "WebRtcRespectOsRoutingTableEnabled"
Hi, if I understand the description of the "WebRtcRespectOsRoutingTableEnabled"-Setting (https://docs.microsoft.com/en-us/deployedge/microsoft-edge-policies#webrtcrespectosroutingtableenabled) correctly, then the setting would need to be set to "Disabled" instead of "Enabled" for the split tunnel scenario to apply. This is not the case here. Or am i wrong?

Description:
"If you disable this policy or don't configure it, WebRTC will not consider the routing table and may make peer to peer connections over any available network.
If you enable this policy, WebRTC will prefer to make peer to peer connections using the indicated network interface for the remote address as indicated in the routing table."

Regards,
Wolfgang